Re-implemented motor wrappers
